Tested functionality:
Get recipes that match userId

Able to delete those recipes individually, or delete all recipes user owns

+++ apps/server/src/recipes/data.ts
@@ -1,43 +1,53 @@
-export function getAllRecipes(env: Env): D1PreparedStatement {
-  return env.DB.prepare('SELECT * FROM recipes LIMIT 100;');
+export function getAllRecipes(env: Env, userId: number): D1PreparedStatement {
+  return env.DB.prepare('SELECT * FROM recipes WHERE userId = ?;').bind(userId);
 }
 
-export function getRecipeById(env: Env, recipeId: number): D1PreparedStatement {
+export function getRecipeById(
+  env: Env,
+  userId: number,
+  recipeId: number,
+): D1PreparedStatement {
   return env.DB.prepare(
     `
 SELECT * FROM recipes
-WHERE id = ?
+WHERE id = ? AND userId = ?
 LIMIT 1;
 `,
-  ).bind(recipeId);
+  ).bind(recipeId, userId);
 }
 
 export function deleteRecipeById(
   env: Env,
   recipeId: number,
+  userId: number,
 ): D1PreparedStatement {
   return env.DB.prepare(
     `DELETE FROM recipes
-WHERE id = ?;
+WHERE id = ? AND userId = ?;
 `,
-  ).bind(recipeId);
+  ).bind(recipeId, userId);
 }
 
-export function deleteAllRecipes(env: Env): D1PreparedStatement {
-  return env.DB.prepare('DELETE FROM recipes;');
+export function deleteAllRecipes(
+  env: Env,
+  userId: number,
+): D1PreparedStatement {
+  return env.DB.prepare(`DELETE FROM recipes WHERE userId = ?;
+    `).bind(userId);
 }
 
 export function updateRecipeById(
   env: Env,
   data: {
     recipeId: number;
+    userId: number;
     name: string;
     steps: string;
     servingSize: number;
     ingredients: string;
   },
 ): D1PreparedStatement {
-  const { name, steps, servingSize, ingredients, recipeId } = data;
+  const { name, steps, servingSize, ingredients, recipeId, userId } = data;
   return env.DB.prepare(
     `
   UPDATE recipes 
@@ -45,9 +55,9 @@ SET name = ?,
 steps = ?,
 servingSize = ?,
 ingredients = ?
-WHERE id == ?;
+WHERE id == ? AND userId = ?;
 `,
-  ).bind(name, steps, servingSize, ingredients, recipeId);
+  ).bind(name, steps, servingSize, ingredients, recipeId, userId);
 }
 
 export function createRecipe(
